requirements.txt
Flask>=2.3
gunicorn>=23.0
peewee>=3.16
python-dotenv>=1.0
psycopg2-binary>=2.9
redis>=4.6